+++ tcex/__metadata__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+"""Package version and other metadata"""
+__author__ = 'ThreatConnect (support@threatconnect.com)'
+__author_email__ = 'support@threatconnect.com'
+__description__ = 'ThreatConnect Exchange App Framework'
+__license__ = 'Apache License, Version 2'
+__package_name__ = 'tcex'
+__url__ = 'https://github.com/ThreatConnect-Inc/tcex'
+__version__ = '2.0.0'
+__download_url__ = f'https://github.com/ThreatConnect-Inc/tcex/tarball/{__version__}'