+++ tcex/resources/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Resources module for TcEx Framework"""
-# flake8: noqa
-from .resources import Resources